move index to a right folder
